+++ src/parser/monster/templates/monster.js
@@ -1,11 +1,14 @@
-export async function newNPC(name) {
+import utils from "../../../lib/utils.js";
+
+export async function newNPC(name, isLegacy) {
   const options = {
     temporary: true,
     displaySheet: false,
   };
   const npcClass = await Actor.create({ name, type: "npc" }, options);
   let npc = npcClass.toObject();
-  npc._id = foundry.utils.randomID();
+  // npc._id = foundry.utils.randomID();
+  npc._id = utils.namedIDStub(`${npc.name}${isLegacy ? " Legacy" : ""}`);
   const flags = {
     dnd5e: {},
     monsterMunch: {},